fixes from reviews
